+++ CHANGELOG.md
@@ -8,11 +8,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 Add eslint and lint the entire codebase.
 Add support for Node 18.
+Add `index.js` tests.
 
 ### Changed
 - Split tests into multiple files.
 - Run tests in github actions against multiple node versions. Remove travis.
 - disable npm publish trigger on push to master.
+- Updated documentation in README.md to reflect other smile SDK docs.
 
 ### Removed
 Drop support for Node 10.